Refactor OCCGeometry, OCCViewer, EdgeComponent into modules
