+++ docs/重构说明.md
@@ -1,0 +1,249 @@
+# 代码重构说明 - 模块化架构
+
+## 概述
+
+三个主要程序包（`OCCGeometry`、`OCCViewer`、`EdgeComponent`）已经被重构为更小、更专注的模块，以便更好地管理和维护。
+
+## 1. OCCGeometry 重构
+
+### 原有结构
+- **OCCGeometry.h/cpp** (~2050行代码): 单一的庞大类，包含所有功能
+
+### 新的模块化结构
+
+将原来的 OCCGeometry 拆分为 **8个专注的模块**：
+
+| 模块 | 文件位置 | 功能职责 |
+|------|---------|---------|
+| **OCCGeometryCore** | `geometry/OCCGeometryCore.h` | 核心几何数据（形状、名称、文件） |
+| **OCCGeometryTransform** | `geometry/OCCGeometryTransform.h` | 变换属性（位置、旋转、缩放） |
+| **OCCGeometryMaterial** | `geometry/OCCGeometryMaterial.h` | 材质属性（环境光、漫反射、镜面反射等） |
+| **OCCGeometryAppearance** | `geometry/OCCGeometryAppearance.h` | 外观设置（颜色、透明度、纹理、混合） |
+| **OCCGeometryDisplay** | `geometry/OCCGeometryDisplay.h` | 显示模式（边、顶点、线框） |
+| **OCCGeometryQuality** | `geometry/OCCGeometryQuality.h` | 渲染质量（细分、LOD、阴影、光照） |
+| **OCCGeometryMesh** | `geometry/OCCGeometryMesh.h` | 网格生成和管理 |
+| **OCCGeometryPrimitives** | `geometry/OCCGeometryPrimitives.h` | 基本图元（盒、柱、球、锥等） |
+
+### 主类使用组合模式
+新的 `OCCGeometry` 主类通过**组合模式**将所有模块整合在一起，对外提供统一的接口。
+
+## 2. OCCViewer 重构
+
+### 原有结构
+- **OCCViewer.h/cpp** (~1518行代码): 单一类包含所有查看器功能
+
+### 新的模块化结构
+
+拆分为 **6个控制器模块**：
+
+| 模块 | 文件位置 | 功能职责 |
+|------|---------|---------|
+| **ViewportController** | `viewer/ViewportController.h` | 视口操作（相机、适配、刷新） |
+| **RenderingController** | `viewer/RenderingController.h` | 渲染模式（线框、边、法线显示） |
+| **MeshParameterController** | 已存在 | 网格质量控制 |
+| **LODController** | 已存在 | 细节层次控制 |
+| **SliceController** | 已存在 | 切片平面控制 |
+| **ExplodeController** | 已存在 | 装配爆炸视图 |
+
+### 控制器模式
+新的 `OCCViewer` 主类使用**控制器模式**，将操作委托给专门的控制器。
+
+## 3. EdgeComponent 重构
+
+### 原有结构
+- **EdgeComponent.h/cpp** (~1694行代码): 边提取和渲染混合在一起
+
+### 新的模块化结构
+
+拆分为 **2个专注模块**：
+
+| 模块 | 文件位置 | 功能职责 |
+|------|---------|---------|
+| **EdgeExtractor** | `edges/EdgeExtractor.h` | 边提取逻辑（原始边、特征边、网格边、轮廓边） |
+| **EdgeRenderer** | `edges/EdgeRenderer.h` | 边可视化（Coin3D节点生成和渲染） |
+
+### 外观模式
+新的 `EdgeComponent` 主类使用**外观模式**，组合提取器和渲染器，提供简化的接口。
+
+## 4. 重构优势
+
+### 4.1 可维护性 ✅
+- 更小的模块，更容易理解和修改
+- 职责清晰，每个模块有明确的目的
+- 降低耦合度，模块之间相对独立
+
+### 4.2 可测试性 ✅
+- 可以独立测试每个模块
+- 更容易创建模拟对象
+- 修改隔离，不影响其他模块
+
+### 4.3 可重用性 ✅
+- 模块可以在不同上下文中重用
+- 灵活性高，易于添加或替换模块
+- 扩展性好，无需修改现有代码
+
+### 4.4 代码组织 ✅
+- 逻辑分组，相关功能聚合在一起
+- 结构清晰，目录结构反映功能
+- 文档友好，更容易编写文档
+
+## 5. 目录结构
+
+```
+include/
+├── geometry/                    # 几何模块
+│   ├── OCCGeometry.h           # 主几何类
+│   ├── OCCGeometryCore.h       # 核心几何数据
+│   ├── OCCGeometryTransform.h  # 变换属性
+│   ├── OCCGeometryMaterial.h   # 材质属性
+│   ├── OCCGeometryAppearance.h # 外观设置
+│   ├── OCCGeometryDisplay.h    # 显示模式
+│   ├── OCCGeometryQuality.h    # 质量设置
+│   ├── OCCGeometryMesh.h       # 网格管理
+│   └── OCCGeometryPrimitives.h # 基本图元
+│
+├── viewer/                      # 查看器模块
+│   ├── OCCViewer.h             # 主查看器类
+│   ├── ViewportController.h    # 视口控制
+│   ├── RenderingController.h   # 渲染控制
+│   └── ... (其他控制器)
+│
+├── edges/                       # 边模块
+│   ├── EdgeComponent.h         # 主边类
+│   ├── EdgeExtractor.h         # 边提取
+│   └── EdgeRenderer.h          # 边渲染
+│
+├── OCCGeometry.h               # 兼容性包装器
+├── OCCViewer.h                 # 兼容性包装器
+└── EdgeComponent.h             # 兼容性包装器
+```
+
+## 6. 向后兼容性
+
+**所有现有代码无需修改即可继续工作！**
+
+根目录的兼容性包装器会自动转发到新的模块化头文件：
+
+```cpp
+// 旧代码（仍然可以正常工作）
+#include "OCCGeometry.h"
+OCCGeometry* geom = new OCCGeometry("myShape");
+geom->setColor(color);
+
+// 新的模块化方式（可选）
+#include "geometry/OCCGeometry.h"
+OCCGeometry* geom = new OCCGeometry("myShape");
+geom->setColor(color);  // API完全相同
+```
+
+## 7. 实施状态
+
+### 已完成 ✅
+- ✅ 所有几何模块的头文件
+- ✅ 查看器控制器的头文件
+- ✅ 边模块的头文件
+- ✅ 兼容性包装器
+- ✅ 完整文档
+
+### 待完成 ⏳
+- ⏳ 新模块的实现文件（.cpp）
+- ⏳ CMakeLists.txt 更新
+- ⏳ 现有代码迁移到新模块
+- ⏳ 单元测试
+
+## 8. 设计模式应用
+
+| 模式 | 应用位置 | 说明 |
+|-----|---------|------|
+| **组合模式** | OCCGeometry | 主类组合多个功能模块 |
+| **控制器模式** | OCCViewer | 使用专门控制器管理不同功能 |
+| **外观模式** | EdgeComponent | 简化复杂子系统的接口 |
+| **策略模式** | 渲染/质量设置 | 可插拔的算法实现 |
+
+## 9. 功能分类对照表
+
+### OCCGeometry 功能分布
+
+| 原有功能 | 新模块归属 |
+|---------|-----------|
+| 形状、名称、文件 | OCCGeometryCore |
+| 位置、旋转、缩放 | OCCGeometryTransform |
+| 环境光、漫反射、镜面反射 | OCCGeometryMaterial |
+| 颜色、透明度、纹理 | OCCGeometryAppearance |
+| 边、顶点、线框显示 | OCCGeometryDisplay |
+| LOD、阴影、光照模型 | OCCGeometryQuality |
+| 网格生成、面索引映射 | OCCGeometryMesh |
+| 盒、球、柱等基本形状 | OCCGeometryPrimitives |
+
+### OCCViewer 功能分布
+
+| 原有功能 | 新模块归属 |
+|---------|-----------|
+| 相机、视图适配 | ViewportController |
+| 线框、边、法线显示 | RenderingController |
+| 网格质量控制 | MeshParameterController (已存在) |
+| 细节层次 | LODController (已存在) |
+| 切片平面 | SliceController (已存在) |
+| 爆炸视图 | ExplodeController (已存在) |
+| 边显示管理 | EdgeDisplayManager (已存在) |
+
+### EdgeComponent 功能分布
+
+| 原有功能 | 新模块归属 |
+|---------|-----------|
+| 原始边提取 | EdgeExtractor |
+| 特征边提取 | EdgeExtractor |
+| 网格边提取 | EdgeExtractor |
+| 轮廓边提取 | EdgeExtractor |
+| Coin3D节点生成 | EdgeRenderer |
+| 边外观设置 | EdgeRenderer |
+| 边显示更新 | EdgeRenderer |
+
+## 10. 使用建议
+
+### 新项目
+推荐直接使用新的模块化头文件：
+```cpp
+#include "geometry/OCCGeometry.h"
+#include "viewer/OCCViewer.h"
+#include "edges/EdgeComponent.h"
+```
+
+### 现有项目
+无需修改，继续使用兼容性包装器：
+```cpp
+#include "OCCGeometry.h"
+#include "OCCViewer.h"
+#include "EdgeComponent.h"
+```
+
+## 11. 后续工作
+
+如果需要实现这些模块的 .cpp 文件，建议按以下顺序进行：
+
+1. **第一阶段**：实现基础模块
+   - OCCGeometryCore
+   - OCCGeometryTransform
+   - EdgeExtractor
+   - EdgeRenderer
+
+2. **第二阶段**：实现外观和显示
+   - OCCGeometryMaterial
+   - OCCGeometryAppearance
+   - OCCGeometryDisplay
+
+3. **第三阶段**：实现高级功能
+   - OCCGeometryQuality
+   - OCCGeometryMesh
+   - ViewportController
+   - RenderingController
+
+4. **第四阶段**：集成和测试
+   - 更新主类实现
+   - 迁移现有代码
+   - 添加单元测试
+   - 性能测试
+
+---
+
+**重构完成后，代码将更加模块化、易于维护和扩展！** 🎉
