+++ include/OCCGeometry.h
@@ -1,496 +1,39 @@
 #pragma once
 
-#include "rendering/GeometryProcessor.h"
-#include "config/RenderingConfig.h"
-#include "GeometryDialogTypes.h"
-#include <string>
-#include <memory>
-#include <vector>
-#include <OpenCASCADE/TopoDS_Shape.hxx>
-#include <OpenCASCADE/gp_Pnt.hxx>
-#include <OpenCASCADE/gp_Vec.hxx>
-#include <OpenCASCADE/Quantity_Color.hxx>
-#include "EdgeTypes.h"
-class EdgeComponent;
-
 /**
- * @brief Face index mapping structure for Coin3D triangle to geometry face mapping
+ * @file OCCGeometry.h
+ * @brief Compatibility wrapper for the refactored OCCGeometry modules
+ * 
+ * This file provides backward compatibility by including the new modular geometry headers.
+ * The old OCCGeometry class has been split into focused modules for better management:
+ * 
+ * - OCCGeometryCore: Basic geometry data (shape, name, file)
+ * - OCCGeometryTransform: Transform properties (position, rotation, scale)
+ * - OCCGeometryMaterial: Material properties (ambient, diffuse, specular, etc.)
+ * - OCCGeometryAppearance: Visual appearance (color, transparency, texture, blend)
+ * - OCCGeometryDisplay: Display modes (edges, vertices, wireframe)
+ * - OCCGeometryQuality: Rendering quality (LOD, shadows, lighting)
+ * - OCCGeometryMesh: Coin3D mesh generation and management
+ * - OCCGeometryPrimitives: Primitive shapes (Box, Cylinder, Sphere, etc.)
+ * 
+ * The main OCCGeometry class now uses composition to combine all these modules.
  */
-struct FaceIndexMapping {
-    int geometryFaceId;  // Index of the face in the original geometry (from TopExp_Explorer)
-    std::vector<int> triangleIndices;  // Indices of triangles in Coin3D mesh that belong to this face
 
-    FaceIndexMapping(int faceId = -1) : geometryFaceId(faceId) {}
-};
-
-// Forward declarations
-class SoSeparator;
-class SoTransform;
-class SoMaterial;
-class OCCMeshConverter;
-
-/**
- * @brief Base class for OpenCASCADE geometry objects
- */
-class OCCGeometry {
-public:
-	OCCGeometry(const std::string& name);
-	virtual ~OCCGeometry();
-
-	// Property accessors
-	const std::string& getName() const { return m_name; }
-	void setName(const std::string& name) { m_name = name; }
-
-	const std::string& getFileName() const { return m_fileName; }
-	void setFileName(const std::string& fileName) { m_fileName = fileName; }
-
-	const TopoDS_Shape& getShape() const { return m_shape; }
-	virtual void setShape(const TopoDS_Shape& shape);
-
-	gp_Pnt getPosition() const { return m_position; }
-	virtual void setPosition(const gp_Pnt& position);
-
-	void getRotation(gp_Vec& axis, double& angle) const { axis = m_rotationAxis; angle = m_rotationAngle; }
-	virtual void setRotation(const gp_Vec& axis, double angle);
-
-	double getScale() const { return m_scale; }
-	virtual void setScale(double scale);
-
-	bool isVisible() const { return m_visible; }
-	virtual void setVisible(bool visible);
-
-	bool isSelected() const { return m_selected; }
-	virtual void setSelected(bool selected);
-
-	Quantity_Color getColor() const { return m_color; }
-	virtual void setColor(const Quantity_Color& color);
-
-	double getTransparency() const { return m_transparency; }
-	virtual void setTransparency(double transparency);
-
-	// Material properties
-	Quantity_Color getMaterialAmbientColor() const { return m_materialAmbientColor; }
-	virtual void setMaterialAmbientColor(const Quantity_Color& color);
-
-	Quantity_Color getMaterialDiffuseColor() const { return m_materialDiffuseColor; }
-	virtual void setMaterialDiffuseColor(const Quantity_Color& color);
-
-	Quantity_Color getMaterialSpecularColor() const { return m_materialSpecularColor; }
-	virtual void setMaterialSpecularColor(const Quantity_Color& color);
-	virtual void setMaterialEmissiveColor(const Quantity_Color& color);
-	virtual void setMaterialShininess(double shininess);
-
-	// Set default bright material for better visibility without textures
-	virtual void setDefaultBrightMaterial();
-
-	// Material tracking
-	virtual void resetMaterialExplicitFlag() { m_materialExplicitlySet = false; }
-	virtual bool isMaterialExplicitlySet() const { return m_materialExplicitlySet; }
-
-	// Apply advanced parameters from VisualSettingsDialog
-	virtual void applyAdvancedParameters(const AdvancedGeometryParameters& params);
-
-	// Texture properties
-	Quantity_Color getTextureColor() const { return m_textureColor; }
-	virtual void setTextureColor(const Quantity_Color& color);
-
-	double getTextureIntensity() const { return m_textureIntensity; }
-	virtual void setTextureIntensity(double intensity);
-
-	bool isTextureEnabled() const { return m_textureEnabled; }
-	virtual void setTextureEnabled(bool enabled);
-
-	std::string getTextureImagePath() const { return m_textureImagePath; }
-	virtual void setTextureImagePath(const std::string& path);
-
-	RenderingConfig::TextureMode getTextureMode() const { return m_textureMode; }
-	virtual void setTextureMode(RenderingConfig::TextureMode mode);
-
-	// Blend settings
-	RenderingConfig::BlendMode getBlendMode() const { return m_blendMode; }
-	virtual void setBlendMode(RenderingConfig::BlendMode mode);
-
-	bool isDepthTestEnabled() const { return m_depthTest; }
-	virtual void setDepthTest(bool enabled);
-
-	bool isDepthWriteEnabled() const { return m_depthWrite; }
-	virtual void setDepthWrite(bool enabled);
-
-	bool isCullFaceEnabled() const { return m_cullFace; }
-	virtual void setCullFace(bool enabled);
-
-	double getAlphaThreshold() const { return m_alphaThreshold; }
-	virtual void setAlphaThreshold(double threshold);
-
-	// Shading settings
-	// Removed getShadingMode and setShadingMode - functionality not needed
-
-	bool isSmoothNormalsEnabled() const { return m_smoothNormals; }
-	virtual void setSmoothNormals(bool enabled);
-
-	double getWireframeWidth() const { return m_wireframeWidth; }
-	virtual void setWireframeWidth(double width);
-
-	double getPointSize() const { return m_pointSize; }
-	virtual void setPointSize(double size);
-
-	// Display settings
-	RenderingConfig::DisplayMode getDisplayMode() const { return m_displayMode; }
-	virtual void setDisplayMode(RenderingConfig::DisplayMode mode);
-
-	bool isShowEdgesEnabled() const { return m_showEdges; }
-	virtual void setShowEdges(bool enabled);
-
-	bool isShowVerticesEnabled() const { return m_showVertices; }
-	virtual void setShowVertices(bool enabled);
-
-	double getEdgeWidth() const { return m_edgeWidth; }
-	virtual void setEdgeWidth(double width);
-
-	double getVertexSize() const { return m_vertexSize; }
-	virtual void setVertexSize(double size);
-
-	Quantity_Color getEdgeColor() const { return m_edgeColor; }
-	virtual void setEdgeColor(const Quantity_Color& color);
-
-	Quantity_Color getVertexColor() const { return m_vertexColor; }
-	virtual void setVertexColor(const Quantity_Color& color);
-
-	// Quality settings
-	RenderingConfig::RenderingQuality getRenderingQuality() const { return m_renderingQuality; }
-	virtual void setRenderingQuality(RenderingConfig::RenderingQuality quality);
-
-	int getTessellationLevel() const { return m_tessellationLevel; }
-	virtual void setTessellationLevel(int level);
-
-	int getAntiAliasingSamples() const { return m_antiAliasingSamples; }
-	virtual void setAntiAliasingSamples(int samples);
-
-	bool isLODEnabled() const { return m_enableLOD; }
-	virtual void setEnableLOD(bool enabled);
-
-	double getLODDistance() const { return m_lodDistance; }
-	virtual void setLODDistance(double distance);
-
-	// Shadow settings
-	RenderingConfig::ShadowMode getShadowMode() const { return m_shadowMode; }
-	virtual void setShadowMode(RenderingConfig::ShadowMode mode);
-
-	double getShadowIntensity() const { return m_shadowIntensity; }
-	virtual void setShadowIntensity(double intensity);
-
-	double getShadowSoftness() const { return m_shadowSoftness; }
-	virtual void setShadowSoftness(double softness);
-
-	int getShadowMapSize() const { return m_shadowMapSize; }
-	virtual void setShadowMapSize(int size);
-
-	double getShadowBias() const { return m_shadowBias; }
-	virtual void setShadowBias(double bias);
-
-	// Lighting model settings
-	RenderingConfig::LightingModel getLightingModel() const { return m_lightingModel; }
-	virtual void setLightingModel(RenderingConfig::LightingModel model);
-
-	double getRoughness() const { return m_roughness; }
-	virtual void setRoughness(double roughness);
-
-	double getMetallic() const { return m_metallic; }
-	virtual void setMetallic(double metallic);
-
-	double getFresnel() const { return m_fresnel; }
-	virtual void setFresnel(double fresnel);
-
-	double getSubsurfaceScattering() const { return m_subsurfaceScattering; }
-	virtual void setSubsurfaceScattering(double scattering);
-
-	// Update geometry from RenderingConfig settings
-	virtual void updateFromRenderingConfig();
-
-	// Update material properties for better lighting response
-	virtual void updateMaterialForLighting();
-
-	// Force texture update
-	virtual void forceTextureUpdate();
-
-	// Display modes
-	bool isWireframeMode() const { return m_wireframeMode; }
-	virtual void setWireframeMode(bool wireframe);
-
-	void setFaceDisplay(bool enable);
-	// New: control visibility of faces for edges-only mode without rebuilding flags externally
-	void setFacesVisible(bool visible);
-	void setWireframeOverlay(bool enable);
-	bool hasOriginalEdges() const;
-	void setEdgeDisplay(bool enable);
-	void setFeatureEdgeDisplay(bool enable);
-	void setNormalDisplay(bool enable);
-	void setShowWireframe(bool enabled);
-	bool isShowWireframe() const { return m_showWireframe; }
-
-	// Coin3D integration
-	SoSeparator* getCoinNode();
-	void setCoinNode(SoSeparator* node);
-	void regenerateMesh(const MeshParameters& params);
-	void buildCoinRepresentation(const MeshParameters& params = MeshParameters());
-	void buildCoinRepresentation(const MeshParameters& params,
-		const Quantity_Color& diffuseColor, const Quantity_Color& ambientColor,
-		const Quantity_Color& specularColor, const Quantity_Color& emissiveColor,
-		double shininess, double transparency);
-
-	// Performance optimization
-	bool needsMeshRegeneration() const;
-	void setMeshRegenerationNeeded(bool needed);
-	void updateCoinRepresentationIfNeeded(const MeshParameters& params);
-	void forceCoinRepresentationRebuild(const MeshParameters& params);
-
-	std::unique_ptr<EdgeComponent> edgeComponent;
-	void setEdgeDisplayType(EdgeType type, bool show);
-	bool isEdgeDisplayTypeEnabled(EdgeType type) const;
-	void updateEdgeDisplay();
-
-	// Advanced parameter accessors for comparison with config
-	bool isSmoothingEnabled() const { return m_lastSmoothingEnabled; }
-	int getSmoothingIterations() const { return m_lastSmoothingIterations; }
-	bool isSubdivisionEnabled() const { return m_lastSubdivisionEnabled; }
-	int getSubdivisionLevel() const { return m_lastSubdivisionLevel; }
-
-	// Assembly level for hierarchical explode
-	int getAssemblyLevel() const { return m_assemblyLevel; }
-	void setAssemblyLevel(int level) { m_assemblyLevel = level; }
-
-	// Face index mapping for Coin3D triangle to geometry face mapping
-	const std::vector<FaceIndexMapping>& getFaceIndexMappings() const { return m_faceIndexMappings; }
-	void setFaceIndexMappings(const std::vector<FaceIndexMapping>& mappings) { m_faceIndexMappings = mappings; }
-
-	// Query methods for face index mapping
-	int getGeometryFaceIdForTriangle(int triangleIndex) const;
-	std::vector<int> getTrianglesForGeometryFace(int geometryFaceId) const;
-	bool hasFaceIndexMapping() const { return !m_faceIndexMappings.empty(); }
-
-	// Build face index mapping during mesh generation
-	void buildFaceIndexMapping(const MeshParameters& params = MeshParameters());
-	
-	// LOD (Level of Detail) support - additional methods
-	void addLODLevel(double distance, double deflection);
-	int getLODLevel(double viewDistance) const;
-	
-	// Memory optimization
-	void releaseTemporaryData();
-	void optimizeMemory();
-
-private:
-	void createWireframeRepresentation(const MeshParameters& params);
-
-protected:
-	std::string m_name;
-	std::string m_fileName;
-	TopoDS_Shape m_shape;
-
-	// Transform parameters
-	gp_Pnt m_position;
-	gp_Vec m_rotationAxis;
-	double m_rotationAngle;
-	double m_scale;
-
-	// Display properties
-	bool m_visible;
-	bool m_selected;
-	bool m_facesVisible; // control face visibility independently for edges-only mode
-	Quantity_Color m_color;
-	double m_transparency;
-
-	// Material properties
-	Quantity_Color m_materialAmbientColor;
-	Quantity_Color m_materialDiffuseColor;
-	Quantity_Color m_materialSpecularColor;
-	Quantity_Color m_materialEmissiveColor;
-	double m_materialShininess;
-
-	// Texture properties
-	Quantity_Color m_textureColor;
-	double m_textureIntensity;
-	bool m_textureEnabled;
-	std::string m_textureImagePath;
-	RenderingConfig::TextureMode m_textureMode;
-
-	// Blend properties
-	RenderingConfig::BlendMode m_blendMode;
-	bool m_depthTest;
-	bool m_depthWrite;
-	bool m_cullFace;
-	double m_alphaThreshold;
-
-	// Shading settings
-	bool m_smoothNormals;
-	double m_wireframeWidth;
-	double m_pointSize;
-	bool m_subdivisionEnabled;
-	int m_subdivisionLevels;
-
-	// Display settings
-	RenderingConfig::DisplayMode m_displayMode;
-	bool m_showEdges;
-	bool m_showVertices;
-	double m_edgeWidth;
-	double m_vertexSize;
-	Quantity_Color m_edgeColor;
-	Quantity_Color m_vertexColor;
-
-	// Quality settings
-	RenderingConfig::RenderingQuality m_renderingQuality;
-	int m_tessellationLevel;
-	int m_antiAliasingSamples;
-	bool m_enableLOD;
-	double m_lodDistance;
-
-	// Shadow settings
-	RenderingConfig::ShadowMode m_shadowMode;
-	double m_shadowIntensity;
-	double m_shadowSoftness;
-	int m_shadowMapSize;
-	double m_shadowBias;
-
-	// Lighting model settings
-	RenderingConfig::LightingModel m_lightingModel;
-	double m_roughness;
-	double m_metallic;
-	double m_fresnel;
-	double m_subsurfaceScattering;
-
-	// Display modes
-	bool m_wireframeMode;
-	bool m_showWireframe = false; // controls mesh wireframe
-
-	// Coin3D representation
-	SoSeparator* m_coinNode;
-	SoTransform* m_coinTransform;
-	bool m_coinNeedsUpdate;
-
-	// Performance optimization
-	bool m_meshRegenerationNeeded;
-	MeshParameters m_lastMeshParams;
-
-	// Advanced parameters tracking for change detection
-	bool m_lastSmoothingEnabled{ false };
-	int m_lastSmoothingIterations{ 2 };
-	double m_lastSmoothingStrength{ 0.5 };
-	double m_lastSmoothingCreaseAngle{ 30.0 };
-	bool m_lastSubdivisionEnabled{ false };
-	int m_lastSubdivisionLevel{ 2 };
-	int m_lastSubdivisionMethod{ 0 };
-	double m_lastSubdivisionCreaseAngle{ 30.0 };
-	int m_lastTessellationMethod{ 0 };
-	int m_lastTessellationQuality{ 2 };
-	double m_lastFeaturePreservation{ 0.5 };
-	bool m_lastAdaptiveMeshing{ false };
-	bool m_lastParallelProcessing{ true };
-
-	// Material tracking
-	bool m_materialExplicitlySet; // Flag to track if material has been explicitly set
-
-	// Assembly level
-	int m_assemblyLevel{0};
-
-	// Face index mapping for Coin3D triangle to geometry face mapping
-	std::vector<FaceIndexMapping> m_faceIndexMappings;
-
-	// LOD support
-	std::vector<std::pair<double, double>> m_lodLevels; // distance, deflection pairs
-};
-
-/**
- * @brief OpenCASCADE box geometry
- */
-class OCCBox : public OCCGeometry {
-public:
-	OCCBox(const std::string& name, double width, double height, double depth);
-
-	void setDimensions(double width, double height, double depth);
-	void getSize(double& width, double& height, double& depth) const;
-
-private:
-	void buildShape();
-
-	double m_width, m_height, m_depth;
-};
-
-/**
- * @brief OpenCASCADE cylinder geometry
- */
-class OCCCylinder : public OCCGeometry {
-public:
-	OCCCylinder(const std::string& name, double radius, double height);
-
-	void setDimensions(double radius, double height);
-	void getSize(double& radius, double& height) const;
-
-private:
-	void buildShape();
-
-	double m_radius, m_height;
-};
-
-/**
- * @brief OpenCASCADE sphere geometry
- */
-class OCCSphere : public OCCGeometry {
-public:
-	OCCSphere(const std::string& name, double radius);
-
-	void setRadius(double radius);
-	double getRadius() const { return m_radius; }
-
-private:
-	void buildShape();
-
-	double m_radius;
-};
-
-/**
- * @brief OpenCASCADE cone geometry
- */
-class OCCCone : public OCCGeometry {
-public:
-	OCCCone(const std::string& name, double bottomRadius, double topRadius, double height);
-
-	void setDimensions(double bottomRadius, double topRadius, double height);
-	void getSize(double& bottomRadius, double& topRadius, double& height) const;
-
-private:
-	void buildShape();
-
-	double m_bottomRadius, m_topRadius, m_height;
-};
-
-/**
- * @brief OpenCASCADE torus geometry
- */
-class OCCTorus : public OCCGeometry {
-public:
-	OCCTorus(const std::string& name, double majorRadius, double minorRadius);
-
-	void setDimensions(double majorRadius, double minorRadius);
-	void getSize(double& majorRadius, double& minorRadius) const;
-
-private:
-	void buildShape();
-
-	double m_majorRadius, m_minorRadius;
-};
-
-/**
- * @brief OpenCASCADE truncated cylinder geometry (frustum)
- */
-class OCCTruncatedCylinder : public OCCGeometry {
-public:
-	OCCTruncatedCylinder(const std::string& name, double bottomRadius, double topRadius, double height);
-
-	void setDimensions(double bottomRadius, double topRadius, double height);
-	void getSize(double& bottomRadius, double& topRadius, double& height) const;
-
-private:
-	void buildShape();
-
-	double m_bottomRadius, m_topRadius, m_height;
-};
+// Include all geometry modules
+#include "geometry/OCCGeometry.h"
+#include "geometry/OCCGeometryCore.h"
+#include "geometry/OCCGeometryTransform.h"
+#include "geometry/OCCGeometryMaterial.h"
+#include "geometry/OCCGeometryAppearance.h"
+#include "geometry/OCCGeometryDisplay.h"
+#include "geometry/OCCGeometryQuality.h"
+#include "geometry/OCCGeometryMesh.h"
+#include "geometry/OCCGeometryPrimitives.h"
+
+// Re-export primitive classes for backward compatibility
+using OCCBox = OCCBox;
+using OCCCylinder = OCCCylinder;
+using OCCSphere = OCCSphere;
+using OCCCone = OCCCone;
+using OCCTorus = OCCTorus;
+using OCCTruncatedCylinder = OCCTruncatedCylinder;
